requirements.txt
argparse
pillow
requests
moviepy
numpy